requirements.txt
# General
requests>=2.31.0
python-dotenv>=1.0.0
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.0

# Module 1: Hunters
selenium>=4.10.0
boto3>=1.34.0
yt-dlp>=2023.10.0
opencv-python-headless>=4.8.0
youtube-transcript-api>=0.6.0
newspaper3k>=0.2.8
lxml_html_clean>=0.1.0

# Module 2: Brains
torch>=2.1.0
transformers>=4.36.0
pillow>=10.0.0
scikit-learn>=1.3.0
numpy>=1.26.0
sentencepiece>=0.1.99
protobuf>=4.25.0

# Module 3: Oracle
pytrends>=4.9.0
pandas>=2.1.0

# Module 4: Creative
google-genai
diffusers>=0.24.0
accelerate>=0.25.0

# Module 6: Automation
fastapi>=0.109.0
uvicorn>=0.27.0
